refactor(edge-worker): Restructure project with improved module organization
Reorganize source code into core, flow, and queue modules
- Move files into new directory structure
- Update import paths to reflect new module organization
- Remove deprecated FlowTaskRecord type
- Simplify and consolidate type imports
- Update project.json to use recursive copy for source files
