+++ pkgs/edge-worker/src/core/BatchProcessor.ts
