+++ pkgs/edge-worker/src/core/ExecutionController.ts
