+++ pkgs/edge-worker/src/core/Heartbeat.ts
