+++ pkgs/edge-worker/src/core/Logger.ts
