+++ pkgs/edge-worker/src/core/Queries.ts
