+++ pkgs/edge-worker/src/core/Worker.ts
