+++ pkgs/edge-worker/src/core/WorkerState.ts
