+++ pkgs/edge-worker/src/core/types.ts
@@ -38,18 +38,6 @@ export interface IBatchProcessor {
   awaitCompletion(): Promise<void>;
 }
 
-/**
- * Fields are nullable because types in postgres does not allow NOT NULL,
- * but all those values except `message` come from queue table columns,
- * which are explicitely marked as NOT NULL.
- */
-export interface PgmqMessageRecord<TPayload extends Json | null = Json> extends IMessage {
-  read_ct: number;
-  enqueued_at: string;
-  vt: string;
-  message: TPayload;
-}
-
 export type WorkerRow = {
   last_heartbeat_at: string;
   queue_name: string;
