+++ pkgs/edge-worker/src/flow/StepTaskExecutor.ts
@@ -1,8 +1,7 @@
-import type { Flow } from '../../dsl/src/dsl.ts';
-import type { StepTaskRecord, IPgflowAdapter } from './types-flow.ts';
-import type { Json } from './types.ts';
-import type { IExecutor } from './types.ts';
-import { getLogger } from './Logger.ts';
+import type { Flow } from '../../../dsl/src/dsl.ts';
+import type { StepTaskRecord, IPgflowAdapter } from './types.ts';
+import type { Json, IExecutor } from '../core/types.ts';
+import { getLogger } from '../core/Logger.ts';
 
 class AbortError extends Error {
   constructor() {
