+++ pkgs/edge-worker/src/queue/Queue.ts
@@ -1,5 +1,6 @@
 import type postgres from 'postgres';
-import type { Json, PgmqMessageRecord } from './types.ts';
+import type { PgmqMessageRecord } from './types.ts';
+import type { Json } from '../core/types.ts';
 
 export class Queue<TPayload extends Json> {
   constructor(private readonly sql: postgres.Sql, readonly queueName: string) {}
