+++ pkgs/edge-worker/src/queue/createQueueWorker.ts
@@ -1,14 +1,15 @@
-import type { EdgeWorkerConfig } from "./EdgeWorker.ts";
-import { ExecutionController } from "./ExecutionController.ts";
+import type { EdgeWorkerConfig } from "../EdgeWorker.ts";
+import { ExecutionController } from "../core/ExecutionController.ts";
 import { MessageExecutor } from "./MessageExecutor.ts";
-import { Queries } from "./Queries.ts";
+import { Queries } from "../core/Queries.ts";
 import { Queue } from "./Queue.ts";
 import { ReadWithPollPoller } from './ReadWithPollPoller.ts';
-import type { Json, PgmqMessageRecord } from './types.ts';
-import { Worker } from './Worker.ts';
+import type { Json } from '../core/types.ts';
+import type { PgmqMessageRecord } from './types.ts';
+import { Worker } from '../core/Worker.ts';
 import postgres from 'postgres';
-import { WorkerLifecycle } from "./WorkerLifecycle.ts";
-import { BatchProcessor } from "./BatchProcessor.ts";
+import { WorkerLifecycle } from "../core/WorkerLifecycle.ts";
+import { BatchProcessor } from "../core/BatchProcessor.ts";
 
 /**
  * Configuration for the queue worker
