+++ pkgs/edge-worker/tests/integration/creating_queue.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "@std/assert";
-import { createQueueWorker } from '../../src/createQueueWorker.ts';
+import { createQueueWorker } from '../../src/queue/createQueueWorker.ts';
 import { withTransaction } from "../db.ts";
 import { delay } from "@std/async";
